requirements.txt
Django==4.2.21
Pillow
WeasyPrint
xhtml2pdf
reportlab
django-import-export
django-filter
tablib[xlsx]
psycopg2-binary
gunicorn
whitenoise
django-admin-interface
